api/requirements.txt
annotated-types==0.7.0
anyio==4.5.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
fastapi==0.115.0
greenlet==3.1.0
h11==0.14.0
idna==3.10
mysql-connector-python==9.0.0
paramiko==3.5.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
